python/requirements.txt
eclipse-zenoh

